requirements.txt
# 核心依赖
pyyaml>=6.0
requests>=2.28.0
pytz>=2022.1

# 开发依赖
pytest>=7.0.0
pytest-cov>=4.0.0